Add support for spaCy.
Fixes #162

- Previous attempt was to create an AnkiSpacy addon that was a package manager
for installing spacy and its models and notifying other addons. This posed
several issues (mainly in windows). See https://github.com/rteabeault/AnkiSpacy/issues/7
- This solution is essentially a continuation of the @ianki solution here
https://github.com/kaegi/MorphMan/pull/193
- It uses a python executable path to execute spaCy to discover what models are installed.
- It uses the same python path to run a subprocess that listens on stdin and
uses spaCy to parse the passed text.
- An unfortunate side affect of this is there is currently not a good way to
kill the subprocess after a recalc. This may not be an issue in practice but
in the future it may be good to have a open/close for morphemizers. These could
be used to initialize the subprocess and then close it after.
- Created a MorphemizerRegistry that contains all registered morphemizers.
Adding and removing morphemizers fires events. The MorphemizerComboBox listens
to these events to keep itself updated.
- A `fake_aqt` was being added to sys.modules for tests but this was done in
all_tests.py. This meant you could not run the tests individually. Moved
all modifications of sys.modules into fake_aqt and added some additional
sys.modules needed for new tests.